More refactoring on main-frontend components
