+++ services/main-frontend/src/components/page-specific/manage/course-instances/id/emails/NewEmailTemplateForm.tsx
@@ -4,7 +4,7 @@ import { TextField } from "@material-ui/core"
 import React, { useState } from "react"
 import { useTranslation } from "react-i18next"
 
-import Button from "../../shared-module/components/Button"
+import Button from "../../../../../../shared-module/components/Button"
 
 const StyledTextField = styled(TextField)`
   margin: 0.3rem;
