+++ services/main-frontend/src/pages/manage/course-instances/[id]/emails.tsx
@@ -5,7 +5,7 @@ import { useTranslation } from "react-i18next"
 import { useQuery } from "react-query"
 
 import Layout from "../../../../components/Layout"
-import NewEmailTemplateForm from "../../../../components/forms/NewEmailTemplateForm"
+import NewEmailTemplateForm from "../../../../components/page-specific/manage/course-instances/id/emails/NewEmailTemplateForm"
 import {
   fetchCourseInstanceEmailTemplates,
   postNewEmailTemplateForCourseInstance,
