+++ shared-module/src/locales/fi/main-frontend.json
@@ -22,7 +22,6 @@
   "error-loading-organizations": "Virhe organisaatioiden lataamisessa",
   "error-title": "Virhe",
   "error-unknown-in-component": "Määrittelemätön virhe komponentissa {{component}}",
-  "export-points-as-csv": "Lataa pisteet CSV-tiedostona",
   "feedback-given": "Palaute: {{feedback-given}}",
   "selected-text": "Valittu teksti: {{selected-text}}",
   "no-selected-text": "Ei valittua tekstiä",
